block: Do not attempt extended merge requests
Since this is an Android kernel, chances are that the main storage is a
non-rotational storage device, with reasonably quick read speeds. By
performing extended merges, we significantly increase block overhead at
the cost of very little to no gain. Hard disks do gain read performance
with extended merges by taking advantage of the hard disk while it is
still awake, rather than waiting for it to spin up.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: ZulFKennedy <botnancy1@gmail.com>
Signed-off-by: anandhan07 <anandhasayanan007jb@gmail.com>